# Fixed up allocations in JaegerUdpClient.