Before:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  DefaultJob : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT


```
|                    Method | NumberOfBatches | NumberOfSpans |     Mean |   Error |  StdDev | Completed Work Items | Lock Contentions |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------------- |-------------- |---------:|--------:|--------:|---------------------:|-----------------:|---------:|------:|------:|----------:|
| JaegerUdpClient_SendAsync |               1 |         10000 | 100.4 ms | 1.95 ms | 2.74 ms |               0.4000 |                - | 400.0000 |     - |     - |   3.51 MB |

After:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  DefaultJob : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT


```
|                    Method | NumberOfBatches | NumberOfSpans |     Mean |    Error |   StdDev | Completed Work Items | Lock Contentions | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |---------------- |-------------- |---------:|---------:|---------:|---------------------:|-----------------:|------:|------:|------:|----------:|
| JaegerUdpClient_SendAsync |               1 |         10000 | 86.23 ms | 1.737 ms | 3.088 ms |               0.3333 |                - |     - |     - |     - |         - |